# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "For the administration of analytics pipelines.",
        "title": "Analytics-Flow-Engine API",
        "contact": {},
        "license": {
            "name": "Apache-2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "{version}"
    },
    "basePath": "/",
    "paths": {
        "/pipeline": {
            "put": {
                "description": "Updates a pipeline",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline"
                ],
                "summary": "Update a pipeline",
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Starts a pipeline",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline"
                ],
                "summary": "Start a pipeline",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_SENERGY-Platform_analytics-pipeline_lib.Pipeline"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pipeline/{id}": {
            "get": {
                "description": "Gets a single pipeline status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline"
                ],
                "summary": "Get pipeline status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pipeline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.PipelineStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a single pipeline",
                "tags": [
                    "Pipeline"
                ],
                "summary": "Delete pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pipeline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pipelines": {
            "post": {
                "description": "Gets multiple pipelines status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline"
                ],
                "summary": "Get multiple pipelines status",
                "parameters": [
                    {
                        "description": "Pipeline status request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.PipelineStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/lib.PipelineStatus"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_SENERGY-Platform_analytics-pipeline_lib.DownstreamConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "instanceID": {
                    "type": "string"
                },
                "serviceID": {
                    "type": "string"
                }
            }
        },
        "github_com_SENERGY-Platform_analytics-pipeline_lib.InputTopic": {
            "type": "object",
            "properties": {
                "filterType": {
                    "type": "string"
                },
                "filterValue": {
                    "type": "string"
                },
                "filterValue2": {
                    "type": "string"
                },
                "mappings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_SENERGY-Platform_analytics-pipeline_lib.Mapping"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_SENERGY-Platform_analytics-pipeline_lib.Mapping": {
            "type": "object",
            "properties": {
                "dest": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "github_com_SENERGY-Platform_analytics-pipeline_lib.Operator": {
            "type": "object",
            "properties": {
                "applicationId": {
                    "type": "string"
                },
                "config": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "cost": {
                    "type": "integer"
                },
                "deploymentType": {
                    "type": "string"
                },
                "downstream": {
                    "$ref": "#/definitions/github_com_SENERGY-Platform_analytics-pipeline_lib.DownstreamConfig"
                },
                "id": {
                    "type": "string"
                },
                "imageId": {
                    "type": "string"
                },
                "inputSelections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.InputSelection"
                    }
                },
                "inputTopics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_SENERGY-Platform_analytics-pipeline_lib.InputTopic"
                    }
                },
                "name": {
                    "type": "string"
                },
                "operatorId": {
                    "type": "string"
                },
                "outputTopic": {
                    "type": "string"
                },
                "persistData": {
                    "type": "boolean"
                },
                "upstream": {
                    "$ref": "#/definitions/github_com_SENERGY-Platform_analytics-pipeline_lib.UpstreamConfig"
                }
            }
        },
        "github_com_SENERGY-Platform_analytics-pipeline_lib.Pipeline": {
            "type": "object",
            "properties": {
                "baggage": {
                    "description": "Baggage is the OpenTelemetry context the caller that started this pipeline\nsent along, plus the pipeline's own id. The flow engine turns it into pod\nlabels and hands it to every operator, so a log line from an operator can be\ntraced back to, for example, the smart service instance it belongs to.\n\nStored here rather than only applied at deploy time because a pipeline\noutlives the request that created it: the flow engine recreates missing\ndeployments on startup and rebuilds them on every update, and would have\nnowhere to read the context back from.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "consumeAllMessages": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "flowId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "mergeStrategy": {
                    "type": "string"
                },
                "metrics": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "operators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_SENERGY-Platform_analytics-pipeline_lib.Operator"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "windowTime": {
                    "type": "integer"
                }
            }
        },
        "github_com_SENERGY-Platform_analytics-pipeline_lib.UpstreamConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                }
            }
        },
        "lib.InputSelection": {
            "type": "object",
            "properties": {
                "aspectId": {
                    "type": "string"
                },
                "characteristicIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "functionId": {
                    "type": "string"
                },
                "inputName": {
                    "description": "references mapping name",
                    "type": "string"
                },
                "selectableId": {
                    "description": "either device or group. can be used for SNRGY-1172, needed to update devices in group",
                    "type": "string"
                }
            }
        },
        "lib.PipelineStatus": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "running": {
                    "type": "boolean"
                },
                "transitioning": {
                    "type": "boolean"
                }
            }
        },
        "lib.PipelineStatusRequest": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}